agent-sim/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-sim"
version = "0.1.0"
description = "The main simulation application for the ARLA project."
dependencies = [
    # External PyPI packages
    "hydra-core",
    "pydantic",
    "celery",
    "redis",
    "mlflow",
    "sqlalchemy",
    "asyncpg", # Or "aiosqlite" for SQLite
    "imageio",
    "matplotlib",

    # Local monorepo packages
    "agent-core",
    "agent-engine",
    "agent-concurrent",
    "agent-persist"
]

[tool.setuptools.packages.find]
where = ["src"]
